+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added 
+- new classes to `sku-selector` component
+
 ## [3.138.1] - 2020-12-28
 ### Changed
 - Makes search placeholder translatable